添加一部龙泉经或者修改龙泉经时，同步其没有创建的卷
